sua feedback

+++ src/main/java/EventManagement/Event/repository/EventOperatorRepository.java
@@ -3,5 +3,6 @@ package EventManagement.Event.repository;
 import EventManagement.Event.entity.EventOperator;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EventOperatorRepository extends JpaRepository<EventOperator, Integer> {
+public interface EventOperatorRepository extends JpaRepository<EventOperator, Long> {
+
 }
